+++ src/components/burger-constructor/burger-constructor-ingredients/burger-constructor-ingredients.tsx
@@ -3,7 +3,7 @@ import { AppDispatch, RootState } from './../../../services/store';
 import { useDrop } from 'react-dnd';
 import { useSelector, useDispatch } from 'react-redux';
 import { BurgerConstructorCard } from "./../burger-constructor-card/burgerconstructor-card";
-import { addIngredient } from "../../../services/burger-constructor/constructor-ingredients";
+import { addIngredient } from "../../../services/burger-constructor/burger-constructor-ingredients";
 
 import styles from './styles.module.css';
