+++ src/components/burger-ingredients/burger-ingredients.tsx
@@ -4,9 +4,9 @@ import { Tab } from '@ya.praktikum/react-developer-burger-ui-components';
 import { ingredientItemTypes, ingredientItemGroupName } from "../../utils/types";
 import { BurgerIngredientsGroup } from "./burger-ingredients-group/burger-ingredients-group";
 import { Modal } from "../modal/modal";
-import { fetchIngredients } from "../../services/thunks/burgeringredients";
-import { resetSelected } from '../../services/burger-ingredients/selected-ingredient';
-import { setCurrentActiveTab } from '../../services/burger-ingredients/current-activetab';
+import { fetchIngredients } from "../../services/thunks/thunks";
+import { resetSelected } from '../../services/burger-ingredients/burger-ingredients-selected-ingredient';
+import { setCurrentActiveTab } from '../../services/burger-ingredients/burger-ingredients-current-activetab';
 import { IngredientDetails } from "./ingredient-details/ingredient-details";
 import { AppDispatch, RootState } from './../../services/store';
 import { getVisibleGroup, groupBy } from "./utils";
