+++ src/services/burger-constructor/burger-constructor-ingredients.ts
@@ -8,22 +8,20 @@ type burgerConstructorInitialStateType = {
     ingredients: ingredientItem[];
 };
 
-const myInitialState: burgerConstructorInitialStateType = {
+const burgerConstructorInitialState: burgerConstructorInitialStateType = {
     bun: null,
     ingredients: []
 };
 
 export const BurgerConstructorSlice = createSlice({
-    name: 'burger-constructor',
-    initialState: myInitialState,
+    name: 'burger-constructor/ingredients',
+    initialState: burgerConstructorInitialState,
     reducers: {
         addIngredient: (state, action) => {
-            state.ingredients = [...state.ingredients, 
-                {
+            state.ingredients = [...state.ingredients, {
                 ...action.payload.item,
                 uuid: uuidv4()
-                }
-        ];
+            }];
         },
         removeIngredient: (state, action) => {
             state.ingredients = state.ingredients.filter(ingredient => ingredient.uuid !== action.payload.uuid);
