+++ src/services/root-reducer.ts
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
-import { BurgerIngredientsSlice } from './burger-ingredients/ingrediets';
-import { currentActiveTabSlice } from './burger-ingredients/current-activetab';
-import { selectedIngredientSlice } from './burger-ingredients/selected-ingredient';
-import { BurgerConstructorSlice } from './burger-constructor/constructor-ingredients'
-import { BurgerConstructorOrderSlice } from './burger-constructor/order'
+import { BurgerIngredientsSlice } from './burger-ingredients/burger-ingredients-ingrediets';
+import { currentActiveTabSlice } from './burger-ingredients/burger-ingredients-current-activetab';
+import { selectedIngredientSlice } from './burger-ingredients/burger-ingredients-selected-ingredient';
+import { BurgerConstructorSlice } from './burger-constructor/burger-constructor-ingredients'
+import { BurgerConstructorOrderSlice } from './burger-constructor/burger-constructor-order'
 
 export const rootReducer = combineReducers({
     burgerIngredientsIngredient: BurgerIngredientsSlice.reducer,
